fix: 修复BUG，优化部分功能

+++ src/fileSystem/lib/FileList.php
@@ -3,60 +3,17 @@
 
 namespace iflow\fileSystem\lib;
 
+use iflow\Container\implement\annotation\implement\initializer\FileSystem;
 
-class FileList
-{
 
-    public string $ext = '';
-
-    /**
-     * 获取文件夹文件
-     * @param string $dir
-     * @param string $ext
-     * @param bool $traverse
-     * @return array
-     */
-    public function loadFileList(string $dir, string $ext = '', bool $traverse = false) : array
-    {
-        $this->ext = $ext?:$this->ext;
-        if (!$traverse) {
-            return glob($dir . '*' . $this->ext);
-        }
-        return match (is_dir($dir)) {
-            true => $this->loadDirFile($dir),
-            false => glob($dir . '*' . $this->ext),
-            default => []
-        };
-    }
-
-    /**
-     * 遍历文件夹
-     * @param string $dir 文件地址
-     * @param array $fileList | 文件列表
-     * @return array
-     */
-    public function loadDirFile(string $dir, array $fileList = []) : array
-    {
-        $iterator = new \FilesystemIterator($dir. DIRECTORY_SEPARATOR);
-        foreach ($iterator as $file) {
-            if (is_dir($file -> getPathname())) {
-                $fileList[$file -> getBasename()] = $this->loadDirFile($file -> getPathname());
-            } else {
-                $fileList[
-                str_replace($this->ext, '', $file -> getBasename())
-                ] = $file -> getPathname();
-            }
-        }
-        return $fileList;
-    }
+class FileList extends FileSystem {
 
     /**
      * 获取当前目录下所有文件夹
      * @param string $dir
      * @return array
      */
-    public function loadDir(string $dir): array
-    {
+    public function loadDir(string $dir): array {
         // 获取目录
         $iterator = new \FilesystemIterator(trim($dir, DIRECTORY_SEPARATOR). DIRECTORY_SEPARATOR);
         $fileList = [];
@@ -89,9 +46,7 @@ class FileList
                     $ignoreCount += $remove['ignoreCount'];
                     $count += $remove['dirCount'];
                     // 检测文件夹是否为空
-                    if (!(new \FilesystemIterator($path)) -> valid()) {
-                        rmdir($path);
-                    }
+                    if (!(new \FilesystemIterator($path)) -> valid()) rmdir($path);
                 } else {
                     $ignoreCount += 1;
                 }
